Separate controller parameter retrieval from creation
- Move variables and functions used for controller parameter
  retrieval from YAML files from `controller_creation.py` to the
  new module `controller_params.py`.
- Update import statements in relevant files.

This change improves separation of concerns between the parameter
retrieval and the creation of Data-Driven MPC controllers.
